+++ examples/get_userid.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License
 
-"""Fetch details for the given user."""
+"""Get the userid corresponding to the given user email."""
 
 from argparse import ArgumentParser
 import json
@@ -20,10 +20,18 @@ from urllib.request import HTTPError
 from railib import api, config, show
 
 
-def run(userid: str, profile: str):
+def get_userid(ctx, email: str) -> str:
+    rsp = api.list_users(ctx)
+    for item in rsp:
+        if item["email"] == email:
+            return item["id"]
+    return None
+
+
+def run(email: str, profile: str):
     cfg = config.read(profile=profile)
     ctx = api.Context(**cfg)
-    rsp = api.get_user(ctx, userid)
+    rsp = get_userid(ctx, email)
     print(json.dumps(rsp, indent=2))
 
 
@@ -31,9 +39,9 @@ if __name__ == "__main__":
     p = ArgumentParser()
     p.add_argument("-p", "--profile", type=str,
                    help="profile name", default="default")
-    p.add_argument("userid", type=str, nargs=1, help="user id")
+    p.add_argument("email", type=str, help="user email")
     args = p.parse_args()
     try:
-        run(args.userid[0], args.profile)
+        run(args.email, args.profile)
     except HTTPError as e:
         show.http_error(e)